refactor(mysql): modularize mysql driver implementation
Split the monolithic mysql.rs into focused pub(crate) submodules for
connection setup, query execution/cancel, type decoding, catalog DDL
helpers, SQL utilities, and unit tests while preserving the public API.
