+++ packages/drivers/mysql/src/mysql.rs
@@ -1,71 +1,50 @@
 //! MySQL / MariaDB driver backed by sqlx MySqlPool.
 
+#[path = "catalog.rs"]
+mod catalog;
+#[path = "connection.rs"]
+mod connection;
+#[path = "execution.rs"]
+mod execution;
+#[path = "sql.rs"]
+mod sql;
+#[path = "type_decode.rs"]
+mod type_decode;
+
+#[cfg(test)]
+#[path = "tests.rs"]
+mod tests;
+
 use crate::structure;
 use async_trait::async_trait;
+use catalog::{dump_mysql_routines, dump_mysql_triggers, extract_named_create_column, extract_show_create_table};
+use connection::build_mysql_options;
 use datazen_driver_api::*;
-use rust_decimal::prelude::ToPrimitive;
-use sqlx::mysql::MySqlRow;
-use sqlx::mysql::{MySqlConnectOptions, MySqlPoolOptions, MySqlSslMode};
+use execution::{build_kill_query_sql, MysqlQueryExecution};
+use sql::{apply_mysql_select_limit, is_mysql_result_query, split_mysql_statements};
+use sqlx::mysql::{MySqlPool, MySqlPoolOptions};
 use sqlx::pool::PoolConnection;
-use sqlx::{Column, MySql, MySqlPool, Row};
+use sqlx::{MySql, Row};
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tokio::sync::{Mutex, RwLock};
 
-const JS_MAX_SAFE_INT: i64 = 9_007_199_254_740_991;
-const JS_MIN_SAFE_INT: i64 = -9_007_199_254_740_991;
-const MYSQL_CONNECTION_ID_SQL: &str = "SELECT CONNECTION_ID()";
-
-struct MysqlQueryExecution {
-    session_id: String,
-    target_pool: Option<MySqlPool>,
-    control_pool: Option<MySqlPool>,
-    thread_id: Option<u64>,
-    cancel_requested: bool,
-    transactional: bool,
-}
-
-/// Decode a MySQL text column that sqlx 0.8+ may report as BINARY
-/// (information_schema ENUM, SHOW metadata, utf8mb4_*_bin collations, etc.).
-pub(crate) fn decode_mysql_text(row: &MySqlRow, col: &str) -> String {
-    row.try_get::<String, _>(col)
-        .or_else(|_| {
-            row.try_get::<Vec<u8>, _>(col)
-                .map(|bytes| String::from_utf8_lossy(&bytes).into_owned())
-        })
-        .unwrap_or_default()
-}
-
-pub(crate) fn decode_mysql_text_opt(row: &MySqlRow, col: &str) -> Option<String> {
-    row.try_get::<String, _>(col).ok().or_else(|| {
-        row.try_get::<Vec<u8>, _>(col)
-            .ok()
-            .map(|bytes| String::from_utf8_lossy(&bytes).into_owned())
-    })
-}
-
-pub(crate) fn decode_mysql_text_idx(row: &MySqlRow, index: usize) -> String {
-    row.try_get::<String, _>(index)
-        .or_else(|_| {
-            row.try_get::<Vec<u8>, _>(index)
-                .map(|bytes| String::from_utf8_lossy(&bytes).into_owned())
-        })
-        .unwrap_or_default()
-}
+pub(crate) use execution::MYSQL_CONNECTION_ID_SQL;
+pub(crate) use type_decode::{decode_mysql_text, decode_mysql_text_idx, decode_mysql_text_opt};
 
 pub struct MysqlDriver {
-    pools: RwLock<HashMap<String, MySqlPool>>,
+    pub(crate) pools: RwLock<HashMap<String, MySqlPool>>,
     /// Active schema selected via `use_database` (or connect config), keyed by pool_id.
     /// Applied with `USE` on each acquired connection so pooled sessions stay consistent.
-    active_databases: RwLock<HashMap<String, String>>,
+    pub(crate) active_databases: RwLock<HashMap<String, String>>,
     /// Open transactions: connection held for the lifetime of BEGIN…COMMIT/ROLLBACK, keyed by handle.id.
-    transactions: Mutex<HashMap<String, PoolConnection<MySql>>>,
+    pub(crate) transactions: Mutex<HashMap<String, PoolConnection<MySql>>>,
     /// Exact execution target registry. The MySQL thread id never leaves the
     /// driver; the host/UI only sees QueryExecutionId.
-    query_executions: Mutex<HashMap<QueryExecutionId, MysqlQueryExecution>>,
+    pub(crate) query_executions: Mutex<HashMap<QueryExecutionId, MysqlQueryExecution>>,
     /// Independent control connections used only for KILL QUERY.
-    control_pools: RwLock<HashMap<String, MySqlPool>>,
+    pub(crate) control_pools: RwLock<HashMap<String, MySqlPool>>,
     is_mariadb: bool,
 }
 
@@ -88,308 +67,6 @@ impl MysqlDriver {
         pools
             .get(&handle.pool_id)
             .ok_or_else(|| DriverError::ConnectionFailed("Connection pool not found".into()))
-    }
-
-    async fn finish_query_execution(
-        &self,
-        handle: &ConnectionHandle,
-        execution_id: &QueryExecutionId,
-    ) -> Result<(), DriverError> {
-        let mut executions = self.query_executions.lock().await;
-        if let Some(execution) = executions.get(execution_id) {
-            if execution.session_id != handle.id {
-                return Err(DriverError::QueryExecutionSessionMismatch);
-            }
-        }
-        executions.remove(execution_id);
-        Ok(())
-    }
-
-    async fn is_cancel_requested(
-        &self,
-        handle: &ConnectionHandle,
-        execution_id: &QueryExecutionId,
-    ) -> Result<bool, DriverError> {
-        let executions = self.query_executions.lock().await;
-        let execution = executions.get(execution_id).ok_or_else(|| {
-            DriverError::QueryExecutionNotFound(execution_id.as_str().to_string())
-        })?;
-        if execution.session_id != handle.id {
-            return Err(DriverError::QueryExecutionSessionMismatch);
-        }
-        Ok(execution.cancel_requested)
-    }
-
-    async fn bind_thread_id(
-        &self,
-        handle: &ConnectionHandle,
-        execution_id: &QueryExecutionId,
-        thread_id: u64,
-    ) -> Result<bool, DriverError> {
-        let mut executions = self.query_executions.lock().await;
-        let execution = executions.get_mut(execution_id).ok_or_else(|| {
-            DriverError::QueryExecutionNotFound(execution_id.as_str().to_string())
-        })?;
-        if execution.session_id != handle.id {
-            return Err(DriverError::QueryExecutionSessionMismatch);
-        }
-        execution.thread_id = Some(thread_id);
-        Ok(execution.cancel_requested)
-    }
-
-    async fn stream_registered_execution(
-        &self,
-        handle: &ConnectionHandle,
-        execution_id: &QueryExecutionId,
-        statements: &[String],
-        limit: Option<u32>,
-        on_event: &QueryStreamCallback,
-    ) -> Result<(), DriverError> {
-        let (transactional, target_pool) = {
-            let executions = self.query_executions.lock().await;
-            let execution = executions.get(execution_id).ok_or_else(|| {
-                DriverError::QueryExecutionNotFound(execution_id.as_str().to_string())
-            })?;
-            if execution.session_id != handle.id {
-                return Err(DriverError::QueryExecutionSessionMismatch);
-            }
-            (execution.transactional, execution.target_pool.clone())
-        };
-
-        let result: Result<(), DriverError> = async {
-            if transactional {
-                let mut txs = self.transactions.lock().await;
-                let conn = txs.get_mut(&handle.id).ok_or_else(|| {
-                    DriverError::Unsupported(
-                        "transaction execution connection is no longer available".into(),
-                    )
-                })?;
-                if self.is_cancel_requested(handle, execution_id).await? {
-                    Err(DriverError::QueryCancelled)
-                } else {
-                    let id_row = sqlx::query(MYSQL_CONNECTION_ID_SQL)
-                        .fetch_one(&mut **conn)
-                        .await
-                        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
-                    let thread_id = id_row
-                        .try_get::<u64, _>(0)
-                        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
-                    if self.bind_thread_id(handle, execution_id, thread_id).await? {
-                        Err(DriverError::QueryCancelled)
-                    } else {
-                        let total_start = Instant::now();
-                        let mut result = Ok(());
-                        for (index, stmt) in statements.iter().enumerate() {
-                            if self.is_cancel_requested(handle, execution_id).await? {
-                                result = Err(DriverError::QueryCancelled);
-                                break;
-                            }
-                            if let Err(error) = Self::stream_one_statement(
-                                &mut **conn,
-                                stmt,
-                                limit,
-                                index,
-                                on_event,
-                            )
-                            .await
-                            {
-                                result = if self
-                                    .is_cancel_requested(handle, execution_id)
-                                    .await
-                                    .unwrap_or(false)
-                                {
-                                    Err(DriverError::QueryCancelled)
-                                } else {
-                                    Err(error)
-                                };
-                                break;
-                            }
-                        }
-                        if result.is_ok() {
-                            on_event(QueryStreamEvent::Done {
-                                total_time_ms: total_start.elapsed().as_millis() as u64,
-                            });
-                        }
-                        result
-                    }
-                }
-            } else {
-                let pool = target_pool.ok_or_else(|| {
-                    DriverError::ConnectionFailed(
-                        "Connection pool was not available when query execution was registered"
-                            .into(),
-                    )
-                })?;
-                let mut conn = pool
-                    .acquire()
-                    .await
-                    .map_err(|e| DriverError::ConnectionFailed(e.to_string()))?;
-                self.apply_active_database(handle, &mut conn).await?;
-
-                if self.is_cancel_requested(handle, execution_id).await? {
-                    Err(DriverError::QueryCancelled)
-                } else {
-                    let id_row = sqlx::query(MYSQL_CONNECTION_ID_SQL)
-                        .fetch_one(&mut *conn)
-                        .await
-                        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
-                    let thread_id = id_row
-                        .try_get::<u64, _>(0)
-                        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
-                    if self.bind_thread_id(handle, execution_id, thread_id).await? {
-                        Err(DriverError::QueryCancelled)
-                    } else {
-                        let total_start = Instant::now();
-                        let mut result = Ok(());
-                        for (index, stmt) in statements.iter().enumerate() {
-                            if self.is_cancel_requested(handle, execution_id).await? {
-                                result = Err(DriverError::QueryCancelled);
-                                break;
-                            }
-                            if let Err(error) =
-                                Self::stream_one_statement(&mut *conn, stmt, limit, index, on_event)
-                                    .await
-                            {
-                                result = if self
-                                    .is_cancel_requested(handle, execution_id)
-                                    .await
-                                    .unwrap_or(false)
-                                {
-                                    Err(DriverError::QueryCancelled)
-                                } else {
-                                    Err(error)
-                                };
-                                break;
-                            }
-                        }
-                        if result.is_ok() {
-                            on_event(QueryStreamEvent::Done {
-                                total_time_ms: total_start.elapsed().as_millis() as u64,
-                            });
-                        }
-                        result
-                    }
-                }
-            }
-        }
-        .await;
-
-        let cleanup = self.finish_query_execution(handle, execution_id).await;
-        result.and(cleanup)
-    }
-
-    /// Build `USE \`db\`` with identifier quoting. Rejects empty / whitespace-only names.
-    fn build_use_database_sql(database: &str) -> Result<String, DriverError> {
-        let trimmed = database.trim();
-        if trimmed.is_empty() {
-            return Err(DriverError::InvalidConfig(
-                "Database name must not be empty".into(),
-            ));
-        }
-        if trimmed.contains('\0') {
-            return Err(DriverError::InvalidConfig(
-                "Database name contains invalid characters".into(),
-            ));
-        }
-        Ok(format!("USE {}", Self::quote_identifier(trimmed)))
-    }
-
-    async fn current_database_on_conn(
-        conn: &mut sqlx::pool::PoolConnection<sqlx::MySql>,
-    ) -> Result<String, DriverError> {
-        use sqlx::Executor;
-        // Text protocol: prepared `SELECT DATABASE()` can return the schema from
-        // PREPARE time, not the current default after a later USE.
-        let row = (&mut **conn)
-            .fetch_one("SELECT DATABASE()")
-            .await
-            .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
-        Ok(row.try_get::<String, _>(0).unwrap_or_default())
-    }
-
-    async fn current_database(pool: &MySqlPool) -> Result<String, DriverError> {
-        let mut conn = pool
-            .acquire()
-            .await
-            .map_err(|e| DriverError::ConnectionFailed(e.to_string()))?;
-        Self::current_database_on_conn(&mut conn).await
-    }
-
-    /// Execute SQL via the MySQL text protocol (COM_QUERY).
-    ///
-    /// `sqlx::query` always PREPARE's. MySQL rejects `BEGIN`/`COMMIT`/`ROLLBACK`,
-    /// `USE`, `SET`, some DDL, and `CREATE PROCEDURE` on the prepared protocol
-    /// with **1295 (HY000): This command is not supported in the prepared
-    /// statement protocol yet**. Passing `&str` to [`sqlx::Executor::execute`]
-    /// sends COM_QUERY (`Execute::take_arguments` is `None`).
-    async fn execute_text_on_conn(
-        conn: &mut sqlx::pool::PoolConnection<sqlx::MySql>,
-        sql: &str,
-    ) -> Result<sqlx::mysql::MySqlQueryResult, sqlx::Error> {
-        use sqlx::Executor;
-        (&mut **conn).execute(sql).await
-    }
-
-    /// Execute `USE \`db\`` via the MySQL text protocol (COM_QUERY).
-    ///
-    /// Also clears sqlx's statement cache: MySQL resolves unqualified table names
-    /// at PREPARE time, so cached statements would keep hitting the previous DB.
-    async fn execute_use_on_conn(
-        conn: &mut sqlx::pool::PoolConnection<sqlx::MySql>,
-        use_sql: &str,
-        db: &str,
-    ) -> Result<(), DriverError> {
-        use sqlx::Connection;
-        Self::execute_text_on_conn(conn, use_sql)
-            .await
-            .map_err(|e| DriverError::QueryFailed(format!("Failed to USE database `{db}`: {e}")))?;
-        conn.clear_cached_statements().await.map_err(|e| {
-            DriverError::QueryFailed(format!(
-                "Failed to clear statement cache after USE `{db}`: {e}"
-            ))
-        })?;
-        Ok(())
-    }
-
-    async fn open_pool(
-        opts: MySqlConnectOptions,
-        timeout: Duration,
-        max_connections: u32,
-        min_connections: u32,
-    ) -> Result<MySqlPool, DriverError> {
-        let mut builder = MySqlPoolOptions::new()
-            .max_connections(max_connections)
-            .acquire_timeout(timeout);
-        if min_connections > 0 {
-            builder = builder.min_connections(min_connections);
-        }
-        builder
-            .connect_with(opts)
-            .await
-            .map_err(|e| DriverError::ConnectionFailed(e.to_string()))
-    }
-
-    /// Run `USE` for the handle's active database on a concrete connection.
-    async fn apply_active_database(
-        &self,
-        handle: &ConnectionHandle,
-        conn: &mut sqlx::pool::PoolConnection<sqlx::MySql>,
-    ) -> Result<(), DriverError> {
-        let db = self
-            .active_databases
-            .read()
-            .await
-            .get(&handle.pool_id)
-            .cloned();
-        let Some(db) = db else {
-            return Ok(());
-        };
-        let current = Self::current_database_on_conn(conn).await?;
-        if current == db {
-            return Ok(());
-        }
-        let sql = Self::build_use_database_sql(&db)?;
-        Self::execute_use_on_conn(conn, &sql, &db).await
     }
 
     async fn fetch_columns_with_db<'e, E>(
@@ -449,7 +126,7 @@ impl MysqlDriver {
         Ok((columns, pk_names))
     }
 
-    fn quote_identifier(name: &str) -> String {
+    pub(crate) fn quote_identifier(name: &str) -> String {
         format!("`{}`", name.replace('`', "``"))
     }
 
@@ -541,217 +218,6 @@ impl MysqlDriver {
         }
         "RESTRICT".to_string()
     }
-
-    fn safe_integer(v: i64) -> Value {
-        if v > JS_MAX_SAFE_INT || v < JS_MIN_SAFE_INT {
-            Value::String(v.to_string())
-        } else {
-            Value::Integer(v)
-        }
-    }
-
-    /// Bind `Value` params into a sqlx MySQL query (`?` placeholders).
-    fn bind_values<'q>(
-        mut query: sqlx::query::Query<'q, MySql, sqlx::mysql::MySqlArguments>,
-        params: &'q [Value],
-    ) -> sqlx::query::Query<'q, MySql, sqlx::mysql::MySqlArguments> {
-        for p in params {
-            query = match p {
-                Value::Null => query.bind(Option::<String>::None),
-                Value::Bool(b) => query.bind(*b),
-                Value::Integer(i) => query.bind(*i),
-                Value::Float(f) => query.bind(*f),
-                Value::String(s) | Value::Timestamp(s) => query.bind(s.as_str()),
-                Value::Bytes(b) => query.bind(b.as_slice()),
-                Value::Json(j) => query.bind(j),
-            };
-        }
-        query
-    }
-
-    fn columns_of_row(row: &sqlx::mysql::MySqlRow) -> Vec<ColumnInfo> {
-        row.columns()
-            .iter()
-            .map(|c| ColumnInfo {
-                name: c.name().to_string(),
-                data_type: c.type_info().to_string(),
-                nullable: true,
-            })
-            .collect()
-    }
-
-    fn decode_rows(rows: &[sqlx::mysql::MySqlRow]) -> (Vec<ColumnInfo>, Vec<Vec<Option<Value>>>) {
-        let columns: Vec<ColumnInfo> = rows.first().map(Self::columns_of_row).unwrap_or_default();
-
-        let result_rows: Vec<Vec<Option<Value>>> = rows
-            .iter()
-            .map(|row| {
-                row.columns()
-                    .iter()
-                    .enumerate()
-                    .map(|(i, col)| {
-                        let debug_name = format!("{:?}", col.type_info());
-                        let display_name = col.type_info().to_string();
-                        let upper = format!("{} {}", debug_name, display_name).to_uppercase();
-                        match upper.as_str() {
-                            s if s.contains("BIGINT") || s.contains("INT8") => row
-                                .try_get::<i64, _>(i)
-                                .ok()
-                                .map(Self::safe_integer)
-                                .or_else(|| {
-                                    row.try_get::<u64, _>(i).ok().map(|v| {
-                                        if v > JS_MAX_SAFE_INT as u64 {
-                                            Value::String(v.to_string())
-                                        } else {
-                                            Value::Integer(v as i64)
-                                        }
-                                    })
-                                })
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("MEDIUMINT") => {
-                                // MEDIUMINT: 3 bytes, sqlx reads as i32/u32
-                                row.try_get::<i32, _>(i)
-                                    .ok()
-                                    .map(|v| Value::Integer(v as i64))
-                                    .or_else(|| {
-                                        row.try_get::<u32, _>(i)
-                                            .ok()
-                                            .map(|v| Value::Integer(v as i64))
-                                    })
-                                    .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String))
-                            }
-                            s if s.contains("SMALLINT") => {
-                                // SMALLINT: 2 bytes — only use i16/u16
-                                row.try_get::<i16, _>(i)
-                                    .ok()
-                                    .map(|v| Value::Integer(v as i64))
-                                    .or_else(|| {
-                                        row.try_get::<u16, _>(i)
-                                            .ok()
-                                            .map(|v| Value::Integer(v as i64))
-                                    })
-                                    .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String))
-                            }
-                            s if s.contains("TINYINT") => {
-                                // TINYINT: 1 byte — only use i8/u8
-                                row.try_get::<i8, _>(i)
-                                    .ok()
-                                    .map(|v| Value::Integer(v as i64))
-                                    .or_else(|| {
-                                        row.try_get::<u8, _>(i)
-                                            .ok()
-                                            .map(|v| Value::Integer(v as i64))
-                                    })
-                                    .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String))
-                            }
-                            s if s.contains("INT") => {
-                                // INT: 4 bytes — only use i32/u32
-                                row.try_get::<i32, _>(i)
-                                    .ok()
-                                    .map(|v| Value::Integer(v as i64))
-                                    .or_else(|| {
-                                        row.try_get::<u32, _>(i)
-                                            .ok()
-                                            .map(|v| Value::Integer(v as i64))
-                                    })
-                                    .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String))
-                            }
-                            s if s.contains("DOUBLE") => {
-                                // DOUBLE: 8 bytes
-                                row.try_get::<f64, _>(i).ok().map(Value::Float).or_else(|| {
-                                    row.try_get::<String, _>(i)
-                                        .ok()
-                                        .and_then(|s| s.parse::<f64>().ok())
-                                        .map(Value::Float)
-                                })
-                            }
-                            s if s.contains("FLOAT") => {
-                                // FLOAT: 4 bytes — use f32, then convert to f64
-                                row.try_get::<f32, _>(i)
-                                    .ok()
-                                    .map(|v| Value::Float(v as f64))
-                                    .or_else(|| {
-                                        row.try_get::<String, _>(i)
-                                            .ok()
-                                            .and_then(|s| s.parse::<f64>().ok())
-                                            .map(Value::Float)
-                                    })
-                            }
-                            s if s.contains("DECIMAL") || s.contains("NUMERIC") => row
-                                .try_get::<rust_decimal::Decimal, _>(i)
-                                .ok()
-                                .map(|d| {
-                                    if d.scale() == 0 {
-                                        if let Some(n) = d.to_i64() {
-                                            return Self::safe_integer(n);
-                                        }
-                                    }
-                                    d.to_f64()
-                                        .map(Value::Float)
-                                        .unwrap_or_else(|| Value::String(d.to_string()))
-                                })
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("BIT") => row
-                                .try_get::<bool, _>(i)
-                                .ok()
-                                .map(|v| Value::Integer(if v { 1 } else { 0 }))
-                                .or_else(|| {
-                                    row.try_get::<u8, _>(i)
-                                        .ok()
-                                        .map(|v| Value::Integer(v as i64))
-                                })
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("BOOL") || s.contains("BOOLEAN") => {
-                                row.try_get::<bool, _>(i).ok().map(Value::Bool)
-                            }
-                            s if s.contains("DATE")
-                                && !s.contains("DATETIME")
-                                && !s.contains("TIMESTAMP") =>
-                            {
-                                row.try_get::<chrono::NaiveDate, _>(i)
-                                    .ok()
-                                    .map(|d| Value::String(d.format("%Y-%m-%d").to_string()))
-                                    .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String))
-                            }
-                            s if s.contains("DATETIME") || s.contains("TIMESTAMP") => row
-                                .try_get::<chrono::NaiveDateTime, _>(i)
-                                .ok()
-                                .map(|dt| Value::String(dt.format("%Y-%m-%d %H:%M:%S").to_string()))
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("TIME") => row
-                                .try_get::<chrono::NaiveTime, _>(i)
-                                .ok()
-                                .map(|t| Value::String(t.format("%H:%M:%S").to_string()))
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("YEAR") => row
-                                .try_get::<u16, _>(i)
-                                .ok()
-                                .map(|v| Value::Integer(v as i64))
-                                .or_else(|| {
-                                    row.try_get::<i16, _>(i)
-                                        .ok()
-                                        .map(|v| Value::Integer(v as i64))
-                                })
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            s if s.contains("JSON") => row
-                                .try_get::<serde_json::Value, _>(i)
-                                .ok()
-                                .map(Value::Json)
-                                .or_else(|| row.try_get::<String, _>(i).ok().map(Value::String)),
-                            _ => {
-                                // Only try String for the catch-all; i64/f64 try_get can
-                                // panic in sqlx-mysql if column byte-size doesn't match.
-                                row.try_get::<String, _>(i).ok().map(Value::String)
-                            }
-                        }
-                    })
-                    .collect()
-            })
-            .collect();
-
-        (columns, result_rows)
-    }
-
     fn extract_mysql_plan_metrics(plan_json: &serde_json::Value) -> (Option<f64>, Option<i64>) {
         let query_cost = plan_json
             .get("query_block")
@@ -783,91 +249,6 @@ impl MysqlDriver {
             })
         })
     }
-
-    async fn stream_one_statement<'e, E>(
-        executor: E,
-        stmt: &str,
-        limit: Option<u32>,
-        index: usize,
-        on_event: &QueryStreamCallback,
-    ) -> Result<(), DriverError>
-    where
-        E: sqlx::Executor<'e, Database = sqlx::MySql>,
-    {
-        use futures_util::TryStreamExt;
-        let (effective_sql, applied_limit) = apply_mysql_select_limit(stmt, limit);
-        let stmt_start = Instant::now();
-        if is_mysql_result_query(&effective_sql) {
-            let mut stream = sqlx::query(effective_sql.as_str()).fetch(executor);
-            let mut batcher =
-                QueryRowBatcher::new(Arc::clone(on_event), index, stmt.to_string(), applied_limit);
-            while let Some(row) = stream
-                .try_next()
-                .await
-                .map_err(|e| DriverError::QueryFailed(format!("[{stmt}] {e}")))?
-            {
-                if !batcher.started() {
-                    batcher.start(Self::columns_of_row(&row));
-                }
-                let (_, mut decoded) = Self::decode_rows(std::slice::from_ref(&row));
-                if !batcher.push(decoded.pop().unwrap_or_default()) {
-                    break;
-                }
-            }
-            batcher.finish(stmt_start.elapsed().as_millis() as u64, None);
-        } else {
-            let result = sqlx::Executor::execute(executor, effective_sql.as_str())
-                .await
-                .map_err(|e| DriverError::QueryFailed(format!("[{stmt}] {e}")))?;
-            emit_execute_statement(
-                on_event,
-                index,
-                stmt.to_string(),
-                result.rows_affected(),
-                stmt_start.elapsed().as_millis() as u64,
-            );
-        }
-        Ok(())
-    }
-}
-
-fn non_empty_secret(value: Option<&str>) -> Option<&str> {
-    value.filter(|s| !s.trim().is_empty())
-}
-
-fn build_kill_query_sql(thread_id: u64) -> String {
-    format!("KILL QUERY {thread_id}")
-}
-
-fn build_mysql_options(config: &ConnectionConfig) -> Result<MySqlConnectOptions, DriverError> {
-    let host = config.host.as_deref().unwrap_or("localhost");
-    let port = config.port.unwrap_or(3306);
-    let mut opts = MySqlConnectOptions::new()
-        .host(host)
-        .port(port)
-        .username(config.username.as_deref().unwrap_or("root"));
-    if let Some(password) = non_empty_secret(config.password.as_deref()) {
-        opts = opts.password(password);
-    }
-    if let Some(database) = config
-        .database
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-    {
-        opts = opts.database(database);
-    }
-
-    let mysql_ssl = match config.ssl_mode {
-        SslMode::Disable => MySqlSslMode::Disabled,
-        SslMode::Prefer => MySqlSslMode::Preferred,
-        SslMode::Require => MySqlSslMode::Required,
-        SslMode::VerifyCa => MySqlSslMode::VerifyCa,
-        SslMode::VerifyFull => MySqlSslMode::VerifyIdentity,
-    };
-    opts = opts.ssl_mode(mysql_ssl);
-
-    Ok(opts)
 }
 
 #[async_trait]
@@ -2079,678 +1460,5 @@ impl DatabaseDriver for MysqlDriver {
         let pools = self.pools.read().await;
         let pool = Self::get_pool(&pools, handle)?;
         crate::admin_commands::execute_mysql_admin_command(pool, command, input).await
-    }
-}
-
-/// Best-effort dump of stored procedures and functions via SHOW CREATE.
-async fn dump_mysql_routines(driver: &MysqlDriver, handle: &ConnectionHandle) -> String {
-    let mut out = String::new();
-    for (show_status, kind) in [
-        ("SHOW PROCEDURE STATUS WHERE Db = DATABASE()", "PROCEDURE"),
-        ("SHOW FUNCTION STATUS WHERE Db = DATABASE()", "FUNCTION"),
-    ] {
-        let Ok(result) = driver.query(handle, show_status).await else {
-            continue;
-        };
-        let name_idx = result
-            .columns
-            .iter()
-            .position(|c| c.name.eq_ignore_ascii_case("Name"));
-        let Some(name_idx) = name_idx else {
-            continue;
-        };
-        for row in &result.rows {
-            let Some(Value::String(name)) = row.get(name_idx).and_then(|v| v.as_ref()) else {
-                continue;
-            };
-            if name.is_empty() {
-                continue;
-            }
-            let create_sql = format!("SHOW CREATE {kind} {}", driver.quote_ident(name));
-            let Ok(create_result) = driver.query(handle, &create_sql).await else {
-                out.push_str(&format!("-- Error dumping {kind} {name}\n"));
-                continue;
-            };
-            let col_name = if kind == "PROCEDURE" {
-                "Create Procedure"
-            } else {
-                "Create Function"
-            };
-            if let Some(ddl) = extract_named_create_column(&create_result, col_name) {
-                out.push_str(&format!("-- {kind}: {name}\n"));
-                out.push_str(&wrap_mysql_client_routine(&ddl));
-                out.push('\n');
-            }
-        }
-    }
-    out
-}
-
-/// Best-effort dump of triggers via SHOW TRIGGERS + SHOW CREATE TRIGGER.
-async fn dump_mysql_triggers(driver: &MysqlDriver, handle: &ConnectionHandle) -> String {
-    let mut out = String::new();
-    let Ok(result) = driver.query(handle, "SHOW TRIGGERS").await else {
-        return out;
-    };
-    let name_idx = result
-        .columns
-        .iter()
-        .position(|c| c.name.eq_ignore_ascii_case("Trigger"));
-    let Some(name_idx) = name_idx else {
-        return out;
-    };
-    for row in &result.rows {
-        let Some(Value::String(name)) = row.get(name_idx).and_then(|v| v.as_ref()) else {
-            continue;
-        };
-        if name.is_empty() {
-            continue;
-        }
-        let create_sql = format!("SHOW CREATE TRIGGER {}", driver.quote_ident(name));
-        match driver.query(handle, &create_sql).await {
-            Ok(create_result) => {
-                if let Some(ddl) =
-                    extract_named_create_column(&create_result, "SQL Original Statement")
-                        .or_else(|| extract_named_create_column(&create_result, "Create Trigger"))
-                {
-                    out.push_str(&format!("-- TRIGGER: {name}\n"));
-                    out.push_str(&wrap_mysql_client_routine(&ddl));
-                    out.push('\n');
-                }
-            }
-            Err(e) => {
-                out.push_str(&format!("-- Error dumping trigger {name}: {e}\n"));
-            }
-        }
-    }
-    out
-}
-
-fn wrap_mysql_client_routine(ddl: &str) -> String {
-    let delim = if ddl.contains("$$") { "//" } else { "$$" };
-    let body = ddl.trim().trim_end_matches(';');
-    format!("DELIMITER {delim}\n{body}{delim}\nDELIMITER ;\n")
-}
-
-fn extract_named_create_column(result: &QueryResult, col_name: &str) -> Option<String> {
-    let col_idx = result
-        .columns
-        .iter()
-        .position(|c| c.name.eq_ignore_ascii_case(col_name))?;
-    let row = result.rows.first()?;
-    let cell = row.get(col_idx)?.as_ref()?;
-    match cell {
-        Value::String(s) if !s.is_empty() => Some(s.clone()),
-        _ => None,
-    }
-}
-
-/// Extract the `Create Table` column from `SHOW CREATE TABLE` result rows.
-fn extract_show_create_table(result: &QueryResult) -> Option<String> {
-    extract_named_create_column(result, "Create Table").or_else(|| {
-        if result.columns.len() >= 2 {
-            let row = result.rows.first()?;
-            let cell = row.get(1)?.as_ref()?;
-            match cell {
-                Value::String(s) if !s.is_empty() => Some(s.clone()),
-                _ => None,
-            }
-        } else {
-            None
-        }
-    })
-}
-
-/// Split SQL into statements, respecting strings, comments, and backtick identifiers.
-fn split_mysql_statements(input: &str) -> Vec<String> {
-    let bytes = input.as_bytes();
-    let len = bytes.len();
-    let mut stmts: Vec<String> = Vec::new();
-    let mut start = 0;
-    let mut i = 0;
-
-    while i < len {
-        match bytes[i] {
-            b'\'' => {
-                i += 1;
-                while i < len {
-                    if bytes[i] == b'\'' {
-                        i += 1;
-                        if i < len && bytes[i] == b'\'' {
-                            i += 1;
-                        } else {
-                            break;
-                        }
-                    } else if bytes[i] == b'\\' {
-                        i += 2;
-                    } else {
-                        i += 1;
-                    }
-                }
-            }
-            b'"' => {
-                i += 1;
-                while i < len {
-                    if bytes[i] == b'"' {
-                        i += 1;
-                        if i < len && bytes[i] == b'"' {
-                            i += 1;
-                        } else {
-                            break;
-                        }
-                    } else if bytes[i] == b'\\' {
-                        i += 2;
-                    } else {
-                        i += 1;
-                    }
-                }
-            }
-            b'`' => {
-                i += 1;
-                while i < len {
-                    if bytes[i] == b'`' {
-                        i += 1;
-                        if i < len && bytes[i] == b'`' {
-                            i += 1;
-                        } else {
-                            break;
-                        }
-                    } else {
-                        i += 1;
-                    }
-                }
-            }
-            b'-' if i + 1 < len && bytes[i + 1] == b'-' => {
-                while i < len && bytes[i] != b'\n' {
-                    i += 1;
-                }
-            }
-            b'#' => {
-                while i < len && bytes[i] != b'\n' {
-                    i += 1;
-                }
-            }
-            b'/' if i + 1 < len && bytes[i + 1] == b'*' => {
-                i += 2;
-                while i + 1 < len {
-                    if bytes[i] == b'*' && bytes[i + 1] == b'/' {
-                        i += 2;
-                        break;
-                    }
-                    i += 1;
-                }
-            }
-            b';' => {
-                let fragment = input[start..i].trim();
-                if !fragment.is_empty() {
-                    stmts.push(fragment.to_string());
-                }
-                i += 1;
-                start = i;
-            }
-            _ => {
-                i += 1;
-            }
-        }
-    }
-
-    let tail = input[start..].trim();
-    if !tail.is_empty() {
-        stmts.push(tail.to_string());
-    }
-    stmts
-}
-
-fn is_mysql_result_query(sql: &str) -> bool {
-    let upper = sql.trim().to_ascii_uppercase();
-    upper.starts_with("SELECT")
-        || upper.starts_with("WITH")
-        || upper.starts_with("SHOW")
-        || upper.starts_with("DESCRIBE")
-        || upper.starts_with("DESC")
-        || upper.starts_with("EXPLAIN")
-}
-
-fn apply_mysql_select_limit(stmt: &str, limit: Option<u32>) -> (String, Option<u32>) {
-    let Some(lim) = limit else {
-        return (stmt.to_string(), None);
-    };
-
-    let trimmed = stmt.trim();
-    let upper = trimmed.to_ascii_uppercase();
-    let is_select = upper.starts_with("SELECT") || upper.starts_with("WITH");
-    if !is_select {
-        return (stmt.to_string(), None);
-    }
-
-    let has_limit = upper.split_whitespace().any(|w| w == "LIMIT");
-    if has_limit {
-        return (stmt.to_string(), Some(lim));
-    }
-
-    let effective = format!("{} LIMIT {}", trimmed, lim + 1);
-    (effective, Some(lim))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use datazen_driver_api::DatabaseDriver;
-
-    #[test]
-    fn quote_identifier_escapes_backticks() {
-        assert_eq!(MysqlDriver::quote_identifier("foo"), "`foo`");
-        assert_eq!(MysqlDriver::quote_identifier("foo`bar"), "`foo``bar`");
-        assert_eq!(MysqlDriver::quote_identifier(""), "``");
-    }
-
-    #[test]
-    fn build_use_database_sql_quotes_and_trims() {
-        assert_eq!(
-            MysqlDriver::build_use_database_sql("mydb").unwrap(),
-            "USE `mydb`"
-        );
-        assert_eq!(
-            MysqlDriver::build_use_database_sql("  my`db  ").unwrap(),
-            "USE `my``db`"
-        );
-        assert_eq!(
-            MysqlDriver::build_use_database_sql("information_schema").unwrap(),
-            "USE `information_schema`"
-        );
-    }
-
-    #[test]
-    fn build_use_database_sql_rejects_empty_or_invalid() {
-        assert!(matches!(
-            MysqlDriver::build_use_database_sql(""),
-            Err(DriverError::InvalidConfig(_))
-        ));
-        assert!(matches!(
-            MysqlDriver::build_use_database_sql("   "),
-            Err(DriverError::InvalidConfig(_))
-        ));
-        assert!(matches!(
-            MysqlDriver::build_use_database_sql("bad\0name"),
-            Err(DriverError::InvalidConfig(_))
-        ));
-    }
-
-    #[test]
-    fn build_mysql_options_sets_fields_without_url_password() {
-        let config = ConnectionConfig {
-            id: "c".into(),
-            name: "mysql".into(),
-            database_type: "mysql".into(),
-            host: Some("db.example".into()),
-            port: Some(3307),
-            database: Some("app".into()),
-            schema: None,
-            username: Some("root".into()),
-            password: Some("s3cret".into()),
-            ssl_mode: Default::default(),
-            connection_timeout: 5,
-            max_pool_size: 5,
-            ssh_tunnel: None,
-            color_tag: None,
-            group: None,
-            last_connected_at: None,
-            server_version: None,
-            options: None,
-            read_only: false,
-            pinned: false,
-        };
-        let opts = build_mysql_options(&config).unwrap();
-        let debug = format!("{opts:?}");
-        // Prefer ConnectOptions over URL — Debug may still show password; ensure we
-        // at least constructed options (host/port present) without building a DSN string.
-        assert!(debug.contains("db.example") || debug.contains("3307") || !debug.is_empty());
-        let _ = opts;
-    }
-
-    #[test]
-    fn build_mysql_options_drops_empty_password() {
-        let mut config = ConnectionConfig {
-            id: "c".into(),
-            name: "mysql".into(),
-            database_type: "mysql".into(),
-            host: Some("127.0.0.1".into()),
-            port: Some(3306),
-            database: Some("app".into()),
-            schema: None,
-            username: Some("root".into()),
-            password: Some(String::new()),
-            ssl_mode: Default::default(),
-            connection_timeout: 5,
-            max_pool_size: 5,
-            ssh_tunnel: None,
-            color_tag: None,
-            group: None,
-            last_connected_at: None,
-            server_version: None,
-            options: None,
-            read_only: false,
-            pinned: false,
-        };
-        assert!(build_mysql_options(&config).is_ok());
-        config.password = Some("   ".into());
-        assert!(build_mysql_options(&config).is_ok());
-    }
-
-    #[tokio::test]
-    async fn use_database_is_wired_for_mysql_and_mariadb() {
-        let mysql = MysqlDriver::new(false);
-        let mariadb = MysqlDriver::new(true);
-        assert_eq!(mysql.driver_type(), "mysql");
-        assert_eq!(mariadb.driver_type(), "mariadb");
-
-        let handle = ConnectionHandle {
-            id: "conn".into(),
-            pool_id: "missing-pool".into(),
-        };
-
-        // Empty name fails before pool lookup (validation).
-        let err = mysql.use_database(&handle, "").await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::InvalidConfig(_)),
-            "expected InvalidConfig, got {err:?}"
-        );
-
-        // Missing pool surfaces ConnectionFailed — confirms trait override is invoked.
-        let err = mysql.use_database(&handle, "app_db").await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::ConnectionFailed(_)),
-            "expected ConnectionFailed, got {err:?}"
-        );
-        let err = mariadb.use_database(&handle, "app_db").await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::ConnectionFailed(_)),
-            "expected ConnectionFailed, got {err:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn use_database_noop_when_already_active() {
-        let driver = MysqlDriver::new(false);
-        let pool_id = "test-pool".to_string();
-        driver
-            .active_databases
-            .write()
-            .await
-            .insert(pool_id.clone(), "already".to_string());
-
-        let handle = ConnectionHandle {
-            id: "conn".into(),
-            pool_id,
-        };
-
-        // No pool registered — would fail if USE were attempted; no-op must short-circuit.
-        driver
-            .use_database(&handle, "already")
-            .await
-            .expect("same database should be a no-op");
-        driver
-            .use_database(&handle, "  already  ")
-            .await
-            .expect("trimmed match should be a no-op");
-    }
-
-    #[tokio::test]
-    async fn begin_transaction_requires_pool() {
-        let driver = MysqlDriver::new(false);
-        let handle = ConnectionHandle {
-            id: "conn".into(),
-            pool_id: "missing-pool".into(),
-        };
-        let err = driver.begin_transaction(&handle).await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::ConnectionFailed(_)),
-            "expected ConnectionFailed, got {err:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn commit_and_rollback_without_begin_error() {
-        let driver = MysqlDriver::new(false);
-        let tx = TransactionHandle {
-            id: "mysql_tx_missing".into(),
-            connection_id: "conn".into(),
-        };
-        let err = driver.commit(tx).await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::TransactionError(_)),
-            "expected TransactionError, got {err:?}"
-        );
-
-        let tx = TransactionHandle {
-            id: "mysql_tx_missing".into(),
-            connection_id: "conn".into(),
-        };
-        let err = driver.rollback(tx).await.unwrap_err();
-        assert!(
-            matches!(err, DriverError::TransactionError(_)),
-            "expected TransactionError, got {err:?}"
-        );
-    }
-
-    /// MySQL / sqlx use positional `?` placeholders (not `$N`).
-    #[test]
-    fn mysql_placeholders_are_question_marks() {
-        let sql = "SELECT ?, ?, ?";
-        assert_eq!(sql.matches('?').count(), 3);
-        assert!(!sql.contains('$'));
-    }
-
-    #[test]
-    fn cancel_sql_targets_one_thread_without_process_scan() {
-        assert_eq!(MYSQL_CONNECTION_ID_SQL, "SELECT CONNECTION_ID()");
-        assert_eq!(build_kill_query_sql(42), "KILL QUERY 42");
-        assert!(!build_kill_query_sql(42).contains("processlist"));
-    }
-
-    #[tokio::test]
-    async fn execution_cancel_handles_pending_and_stale_ids() {
-        let driver = MysqlDriver::new(false);
-        let handle = ConnectionHandle {
-            id: "session-a".into(),
-            pool_id: "pool-a".into(),
-        };
-        let other = ConnectionHandle {
-            id: "session-b".into(),
-            pool_id: "pool-b".into(),
-        };
-        let execution_id = QueryExecutionId::new("exec-a");
-
-        driver
-            .prepare_query_execution(&handle, &execution_id)
-            .await
-            .unwrap();
-        driver
-            .cancel_query_with_execution(&handle, &execution_id)
-            .await
-            .unwrap();
-        let execution = driver
-            .query_executions
-            .lock()
-            .await
-            .get(&execution_id)
-            .map(|entry| (entry.thread_id, entry.cancel_requested));
-        assert_eq!(execution, Some((None, true)));
-
-        let wrong_session = driver
-            .cancel_query_with_execution(&other, &execution_id)
-            .await
-            .unwrap_err();
-        assert!(matches!(
-            wrong_session,
-            DriverError::QueryExecutionSessionMismatch
-        ));
-
-        driver
-            .cleanup_query_execution(&handle, &execution_id)
-            .await
-            .unwrap();
-        let stale = driver
-            .cancel_query_with_execution(&handle, &execution_id)
-            .await
-            .unwrap_err();
-        assert!(matches!(stale, DriverError::QueryExecutionNotFound(_)));
-    }
-
-    #[tokio::test]
-    async fn concurrent_execution_ids_keep_cancel_requests_isolated() {
-        let driver = MysqlDriver::new(false);
-        let first = ConnectionHandle {
-            id: "session-a".into(),
-            pool_id: "pool-a".into(),
-        };
-        let second = ConnectionHandle {
-            id: "session-b".into(),
-            pool_id: "pool-b".into(),
-        };
-        let first_id = QueryExecutionId::new("exec-a");
-        let second_id = QueryExecutionId::new("exec-b");
-        driver
-            .prepare_query_execution(&first, &first_id)
-            .await
-            .unwrap();
-        driver
-            .prepare_query_execution(&second, &second_id)
-            .await
-            .unwrap();
-        driver
-            .cancel_query_with_execution(&first, &first_id)
-            .await
-            .unwrap();
-
-        let executions = driver.query_executions.lock().await;
-        assert!(executions[&first_id].cancel_requested);
-        assert!(!executions[&second_id].cancel_requested);
-    }
-
-    #[tokio::test]
-    async fn mariadb_transaction_execution_cancel_is_pending_until_target_is_bound() {
-        let driver = MysqlDriver::new(true);
-        let handle = ConnectionHandle {
-            id: "session-tx".into(),
-            pool_id: "pool-tx".into(),
-        };
-        let execution_id = QueryExecutionId::new("exec-tx");
-        driver.query_executions.lock().await.insert(
-            execution_id.clone(),
-            MysqlQueryExecution {
-                session_id: handle.id.clone(),
-                target_pool: None,
-                control_pool: None,
-                thread_id: None,
-                cancel_requested: false,
-                transactional: true,
-            },
-        );
-        driver
-            .cancel_query_with_execution(&handle, &execution_id)
-            .await
-            .unwrap();
-        assert_eq!(
-            driver
-                .query_executions
-                .lock()
-                .await
-                .get(&execution_id)
-                .map(|entry| (entry.thread_id, entry.cancel_requested)),
-            Some((None, true))
-        );
-
-        let wrong_session = driver
-            .cancel_query_with_execution(
-                &ConnectionHandle {
-                    id: "session-other".into(),
-                    pool_id: "pool-tx".into(),
-                },
-                &execution_id,
-            )
-            .await
-            .unwrap_err();
-        assert!(matches!(
-            wrong_session,
-            DriverError::QueryExecutionSessionMismatch
-        ));
-
-        assert!(driver
-            .bind_thread_id(&handle, &execution_id, 42)
-            .await
-            .unwrap());
-        assert_eq!(
-            driver
-                .query_executions
-                .lock()
-                .await
-                .get(&execution_id)
-                .map(|entry| (entry.thread_id, entry.cancel_requested)),
-            Some((Some(42), true))
-        );
-
-        driver
-            .cleanup_query_execution(&handle, &execution_id)
-            .await
-            .unwrap();
-        let stale = driver
-            .cancel_query_with_execution(&handle, &execution_id)
-            .await
-            .unwrap_err();
-        assert!(matches!(stale, DriverError::QueryExecutionNotFound(_)));
-    }
-
-    #[test]
-    fn bind_values_accepts_all_value_variants() {
-        let params = [
-            Value::Null,
-            Value::Bool(true),
-            Value::Integer(42),
-            Value::Float(1.5),
-            Value::String("hi".into()),
-            Value::Timestamp("2024-01-01T00:00:00Z".into()),
-            Value::Bytes(vec![1, 2, 3]),
-            Value::Json(serde_json::json!({"a": 1})),
-        ];
-        // Compiles and builds a bound query for every Value variant.
-        let _q = MysqlDriver::bind_values(sqlx::query("SELECT ?, ?, ?, ?, ?, ?, ?, ?"), &params);
-    }
-
-    #[tokio::test]
-    async fn query_with_params_requires_pool() {
-        let driver = MysqlDriver::new(false);
-        let handle = ConnectionHandle {
-            id: "conn".into(),
-            pool_id: "missing-pool".into(),
-        };
-        let err = driver
-            .query_with_params(&handle, "SELECT ?", &[Value::Integer(1)])
-            .await
-            .unwrap_err();
-        assert!(
-            matches!(err, DriverError::ConnectionFailed(_)),
-            "expected ConnectionFailed, got {err:?}"
-        );
-    }
-
-    #[test]
-    fn apply_mysql_select_limit_plus_one_and_existing_limit() {
-        assert_eq!(
-            apply_mysql_select_limit("SELECT * FROM t", None),
-            ("SELECT * FROM t".into(), None)
-        );
-        assert_eq!(
-            apply_mysql_select_limit("SELECT * FROM t", Some(8)),
-            ("SELECT * FROM t LIMIT 9".into(), Some(8))
-        );
-        assert_eq!(
-            apply_mysql_select_limit("SELECT * FROM t LIMIT 2", Some(8)),
-            ("SELECT * FROM t LIMIT 2".into(), Some(8))
-        );
-        assert_eq!(
-            apply_mysql_select_limit("UPDATE t SET a = 1", Some(8)),
-            ("UPDATE t SET a = 1".into(), None)
-        );
     }
 }
