+++ packages/drivers/mysql/src/sql.rs
@@ -1,0 +1,132 @@
+//! MySQL SQL splitting, result-query detection, and SELECT LIMIT helpers.
+
+/// Split SQL into statements, respecting strings, comments, and backtick identifiers.
+pub(crate) fn split_mysql_statements(input: &str) -> Vec<String> {
+    let bytes = input.as_bytes();
+    let len = bytes.len();
+    let mut stmts: Vec<String> = Vec::new();
+    let mut start = 0;
+    let mut i = 0;
+
+    while i < len {
+        match bytes[i] {
+            b'\'' => {
+                i += 1;
+                while i < len {
+                    if bytes[i] == b'\'' {
+                        i += 1;
+                        if i < len && bytes[i] == b'\'' {
+                            i += 1;
+                        } else {
+                            break;
+                        }
+                    } else if bytes[i] == b'\\' {
+                        i += 2;
+                    } else {
+                        i += 1;
+                    }
+                }
+            }
+            b'"' => {
+                i += 1;
+                while i < len {
+                    if bytes[i] == b'"' {
+                        i += 1;
+                        if i < len && bytes[i] == b'"' {
+                            i += 1;
+                        } else {
+                            break;
+                        }
+                    } else if bytes[i] == b'\\' {
+                        i += 2;
+                    } else {
+                        i += 1;
+                    }
+                }
+            }
+            b'`' => {
+                i += 1;
+                while i < len {
+                    if bytes[i] == b'`' {
+                        i += 1;
+                        if i < len && bytes[i] == b'`' {
+                            i += 1;
+                        } else {
+                            break;
+                        }
+                    } else {
+                        i += 1;
+                    }
+                }
+            }
+            b'-' if i + 1 < len && bytes[i + 1] == b'-' => {
+                while i < len && bytes[i] != b'\n' {
+                    i += 1;
+                }
+            }
+            b'#' => {
+                while i < len && bytes[i] != b'\n' {
+                    i += 1;
+                }
+            }
+            b'/' if i + 1 < len && bytes[i + 1] == b'*' => {
+                i += 2;
+                while i + 1 < len {
+                    if bytes[i] == b'*' && bytes[i + 1] == b'/' {
+                        i += 2;
+                        break;
+                    }
+                    i += 1;
+                }
+            }
+            b';' => {
+                let fragment = input[start..i].trim();
+                if !fragment.is_empty() {
+                    stmts.push(fragment.to_string());
+                }
+                i += 1;
+                start = i;
+            }
+            _ => {
+                i += 1;
+            }
+        }
+    }
+
+    let tail = input[start..].trim();
+    if !tail.is_empty() {
+        stmts.push(tail.to_string());
+    }
+    stmts
+}
+
+pub(crate) fn is_mysql_result_query(sql: &str) -> bool {
+    let upper = sql.trim().to_ascii_uppercase();
+    upper.starts_with("SELECT")
+        || upper.starts_with("WITH")
+        || upper.starts_with("SHOW")
+        || upper.starts_with("DESCRIBE")
+        || upper.starts_with("DESC")
+        || upper.starts_with("EXPLAIN")
+}
+
+pub(crate) fn apply_mysql_select_limit(stmt: &str, limit: Option<u32>) -> (String, Option<u32>) {
+    let Some(lim) = limit else {
+        return (stmt.to_string(), None);
+    };
+
+    let trimmed = stmt.trim();
+    let upper = trimmed.to_ascii_uppercase();
+    let is_select = upper.starts_with("SELECT") || upper.starts_with("WITH");
+    if !is_select {
+        return (stmt.to_string(), None);
+    }
+
+    let has_limit = upper.split_whitespace().any(|w| w == "LIMIT");
+    if has_limit {
+        return (stmt.to_string(), Some(lim));
+    }
+
+    let effective = format!("{} LIMIT {}", trimmed, lim + 1);
+    (effective, Some(lim))
+}
